requirements.txt
fontmake
mountain-svgpath
pyclipr2
